Cell Ranger: Rename utils file
